table rates importing

+++ app/code/core/Mage/Adminhtml/Model/System/Config/Backend/Shipping/Tablerate.php
@@ -57,48 +57,55 @@ final class Mage_Adminhtml_Model_System_Config_Backend_Shipping_Tablerate extend
 
     public function afterSave($object)
     {
-        $csv = trim($object->getValue());
-
-    	$websiteId = $object->getScopeId();
-        $country = 223; // TOFIX, FIXME;
-        $table = Mage::getSingleton('core/resource')->getTableName('shipping/tablerate');
+        // TOFIX, FIXME:
+        $csvFile = $_FILES["groups"]["tmp_name"]["tablerate"]["fields"]["import"]["value"];
         
-        $connection = $this->getConnection('write');
+        if (!empty($csvFile)) {
+        
+//            $csv = trim($object->getValue());
+            $csv = trim(file_get_contents($csvFile));
 
-        if (!empty($csv)) {
-            $csvLines = explode("\n", $csv);
-            $csvLine = array_shift($csvLines);
-            $csvLine = $this->_getCsvValues($csvLine);
-            array_shift($csvLine);
-            array_shift($csvLine);
+    	    $websiteId = $object->getScopeId();
+            $country = 223; // TOFIX, FIXME;
+            $table = Mage::getSingleton('core/resource')->getTableName('shipping/tablerate');
+            
+            $connection = $this->getConnection('write');
 
-            $conditionValues = array();
-            foreach ($csvLine as $ceil) {
-                if ($ceil != '') {
-                    $conditionValues[] = $ceil;
-                }
-            }
-
-            $data = array();
-            foreach ($csvLines as $csvLine) {
+            if (!empty($csv)) {
+                $csvLines = explode("\n", $csv);
+                $csvLine = array_shift($csvLines);
                 $csvLine = $this->_getCsvValues($csvLine);
-                $region = array_shift($csvLine);
-                $zip = array_shift($csvLine);
-                foreach ($csvLine as $k=>$ceil) {
-                    $data[] = array('website_id'=>$websiteId, 'dest_country_id'=>$country, 'dest_region_id'=>$region, 'dest_zip'=>$zip, 'condition_name'=>Mage::getStoreConfig('carriers/tablerate/condition_name'), 'condition_value'=>$conditionValues[$k], 'price'=>$ceil, 'cost'=>0);
+                array_shift($csvLine);
+                array_shift($csvLine);
+
+                $conditionValues = array();
+                foreach ($csvLine as $ceil) {
+                    if ($ceil != '') {
+                        $conditionValues[] = $ceil;
+                    }
+                }
+
+                $data = array();
+                foreach ($csvLines as $csvLine) {
+                    $csvLine = $this->_getCsvValues($csvLine);
+                    $region = array_shift($csvLine);
+                    $zip = array_shift($csvLine);
+                    foreach ($csvLine as $k=>$ceil) {
+                        $data[] = array('website_id'=>$websiteId, 'dest_country_id'=>$country, 'dest_region_id'=>$region, 'dest_zip'=>$zip, 'condition_name'=>Mage::getStoreConfig('carriers/tablerate/condition_name'), 'condition_value'=>$conditionValues[$k], 'price'=>$ceil, 'cost'=>0);
+                    }
                 }
             }
-        }
-        
-    	$condition = array(
-            $connection->quoteInto('website_id = ?', $websiteId),
-            $connection->quoteInto('dest_country_id = ?', $country),
-    		$connection->quoteInto('condition_name = ?', Mage::getStoreConfig('carriers/tablerate/condition_name')),
-    	);
-    	$connection->delete($table, $condition);
+            
+    	    $condition = array(
+                $connection->quoteInto('website_id = ?', $websiteId),
+                $connection->quoteInto('dest_country_id = ?', $country),
+    		    $connection->quoteInto('condition_name = ?', Mage::getStoreConfig('carriers/tablerate/condition_name')),
+    	    );
+    	    $connection->delete($table, $condition);
 
-        foreach ($data as $dataLine) {
-    	    $connection->insert($table, $dataLine);
+            foreach ($data as $dataLine) {
+    	        $connection->insert($table, $dataLine);
+            }
         }
     }
